separated the typescript files into multiple, updated tests and paths
